filibustr: Data Utilities for Congressional Research

Provides easy-to-understand and consistent interfaces for accessing data on the U.S. Congress. The functions in 'filibustr' streamline the process for importing data on Congress into R, removing the need to download and work from CSV files and the like. Data sources include 'Voteview' (https://voteview.com/), the U.S. Senate website (https://www.senate.gov/), and more.

See also

Author

Maintainer: Max Feinleib mhfeinleib@gmail.com (ORCID) [copyright holder]