Get replication data from Harbridge-Yong, Volden, and Wiseman (2023)

get_hvw_data() returns replication data from:

Harbridge-Yong, L., Volden, C., & Wiseman, A. E. (2023). The bipartisan path to effective lawmaking. The Journal of Politics, 85(3), 1048–1063. doi:10.1086/723805

Usage

get_hvw_data(chamber, local_path = NULL)

Arguments

chamber

Which chamber to get data for. Options are:

"house", "h", "hr": House data only.

"senate", "s", "sen": Senate data only.

These options are case-insensitive. Any other argument results in an error.

Note: Unlike the Voteview functions, there is no "all" option. The House and Senate data do not have the same number of variables, or the same variable names, so it is not trivial to join the two tables.

You must specify either House or Senate data, since there is no "default" option.

local_path

(Optional) A file path for reading from a local file. If no local_path is specified, will read data from the Harvard Dataverse website.

Value

A tibble.

Details

The replication data is available at the Harvard Dataverse (doi:10.7910/DVN/EARLA4 ).

The House and Senate data come from the files HarbridgeYong_Volden_Wiseman_House_Replication.tab and HarbridgeYong_Volden_Wiseman_Senate_Replication.tab, respectively.

The data spans the 93rd through 114th Congresses (1973-2016).

These datasets have been dedicated to the public domain under CC0 1.0.

Examples

get_hvw_data("senate")
#> Error in httr2::req_perform(httr2::req_error(httr2::req_retry(httr2::req_user_agent(httr2::request(url),     "filibustr R package (https://cran.r-project.org/package=filibustr)"),     max_tries = 3), body = error_body)): HTTP 504 Gateway Timeout.
#> ℹ ERROR 504 when retrieving online data from the Harvard Dataverse website.
if (FALSE) { # interactive() && !is.null(curl::nslookup("dataverse.harvard.edu", error = FALSE))
get_hvw_data("house")
}