Get Legislative Effectiveness Scores data

get_les() returns Legislative Effectiveness Scores data from the Center for Effective Lawmaking.

Usage

get_les(chamber, les_2 = FALSE, local_path = NULL)

Arguments

chamber

Which chamber to get data for. Options are:

"house", "h", "hr": House data only.

"senate", "s", "sen": Senate data only.

These options are case-insensitive. Any other argument results in an error.

Note: Unlike the Voteview functions, there is no "all" option. You must specify either House or Senate data, since there is no "default" option.

There are non-trivial differences between the House and Senate datasets, so take care when joining House and Senate data. Important differences include:

Legislator names are formatted differently. The Senate data has first and last name columns, while the House data has a single thomas_name column.

The year column refers to the first year of the Congress in the House data, but year refers to the preceding election year in the Senate data. Thus, the year for House members is one after that of senators in the same Congress.

les_2

Whether to use LES 2.0 (instead of Classic Legislative Effectiveness Scores). LES 2.0 credits lawmakers when language from their sponsored bills is included in other legislators' bills that become law. LES 2.0 is only available for the 117th Congress. Classic LES is available for the 93rd through 117th Congresses.

local_path

(Optional) A file path for reading from a local file. If no local_path is specified, will read data from the Center for Effective Lawmaking website.

Value

A tibble.

Details

See the Center for Effective Lawmaking website for more information on their data.

The Legislative Effectiveness Score methodology was introduced in:

Volden, C., & Wiseman, A. E. (2014). Legislative effectiveness in the United States Congress: The lawmakers. Cambridge University Press. doi:10.1017/CBO9781139032360

Examples

if (FALSE) { # interactive()
# Classic LES data (93rd-117th Congresses)
get_les("house", les_2 = FALSE)
get_les("senate", les_2 = FALSE)
}
# LES 2.0 (117th Congress)
get_les("house", les_2 = TRUE)
#> # A tibble: 454 × 60
#>    thomas_num thomas_name       icpsr congress  year st_name    cd dem   elected
#>         <int> <chr>             <int>    <int> <int> <fct>   <int> <lgl>   <int>
#>  1      10700 Adams, Alma       21545      117  2021 NC         12 TRUE     2014
#>  2      10701 Aderholt, Robert  29701      117  2021 AL          4 FALSE    1996
#>  3      10702 Aguilar, Pete     21506      117  2021 CA         31 TRUE     2014
#>  4      10703 Allen, Rick       21516      117  2021 GA         12 FALSE    2014
#>  5      10704 Allred, Colin     21900      117  2021 TX         32 TRUE     2018
#>  6      10705 Amodei, Mark      21196      117  2021 NV          2 FALSE    2011
#>  7      10706 Armstrong, Kelly  21901      117  2021 ND          1 FALSE    2018
#>  8      10707 Arrington, Jodey  21700      117  2021 TX         19 FALSE    2016
#>  9      10708 Auchincloss, Jake 22100      117  2021 MA          4 TRUE     2020
#> 10      10709 Axne, Cindy       21902      117  2021 IA          3 TRUE     2018
#> # ℹ 444 more rows
#> # ℹ 51 more variables: female <lgl>, votepct <int>, dwnom1 <dbl>, dwnom2 <dbl>,
#> #   deleg_size <int>, speaker <lgl>, subchr <lgl>, afam <lgl>, latino <lgl>,
#> #   votepct_sq <int>, power <lgl>, chair <lgl>, state_leg <lgl>,
#> #   state_leg_prof <dbl>, majority <lgl>, maj_leader <lgl>, min_leader <lgl>,
#> #   meddist <dbl>, majdist <dbl>, leslag <dbl>, freshman <lgl>,
#> #   seniority <int>, party_code <int>, bioname <chr>, bioguide_id <chr>, …
get_les("senate", les_2 = TRUE)
#> # A tibble: 100 × 60
#>    last     first state congress cgnum icpsr  year dem   majority elected female
#>    <chr>    <chr> <fct>    <int> <int> <int> <int> <lgl> <lgl>      <int> <lgl> 
#>  1 Blunt    Roy   MO         117  2461 29735  2020 FALSE FALSE       2010 FALSE 
#>  2 Brown    Sher… OH         117  2465 29389  2020 TRUE  TRUE        2006 FALSE 
#>  3 Burr     Rich… NC         117  2466 29548  2020 FALSE FALSE       2004 FALSE 
#>  4 Baldwin  Tammy WI         117  2456 29940  2020 TRUE  TRUE        2012 TRUE  
#>  5 Boozman  John  AR         117  2463 20101  2020 FALSE FALSE       2010 FALSE 
#>  6 Blackbu… Mars… TN         117  2459 20351  2020 FALSE FALSE       2018 TRUE  
#>  7 Barrasso John  WY         117  2457 40707  2020 FALSE FALSE       2007 FALSE 
#>  8 Bennet   Mich… CO         117  2458 40910  2020 TRUE  TRUE        2009 FALSE 
#>  9 Blument… Rich… CT         117  2460 41101  2020 TRUE  TRUE        2010 FALSE 
#> 10 Booker   Cory  NJ         117  2462 41308  2020 TRUE  TRUE        2013 FALSE 
#> # ℹ 90 more rows
#> # ℹ 49 more variables: afam <lgl>, latino <lgl>, votepct <int>, chair <lgl>,
#> #   subchr <lgl>, seniority <int>, state_leg <lgl>, state_leg_prof <dbl>,
#> #   maj_leader <lgl>, min_leader <lgl>, votepct_sq <int>, lagles <dbl>,
#> #   power <lgl>, freshman <lgl>, sensq <int>, deleg_size <int>,
#> #   party_code <int>, bioname <chr>, bioguide_id <chr>, born <int>, died <int>,
#> #   dwnom1 <dbl>, dwnom2 <dbl>, meddist <dbl>, majdist <dbl>, cbill2 <int>, …